refactor: move types file
