+++ providers/limiter_provider.ts
@@ -11,7 +11,7 @@
 /// <reference types="@adonisjs/redis/redis_provider" />
 
 import { ApplicationService } from '@adonisjs/core/types'
-import { LimiterService } from '../src/types/main.js'
+import { LimiterService } from '../src/types.js'
 import { LimiterManager } from '../src/limiter_manager.js'
 import { configProvider } from '@adonisjs/core'
 import { RuntimeException } from '@poppinss/utils'
