+++ services/index.ts
@@ -8,7 +8,7 @@
  */
 
 import app from '@adonisjs/core/services/app'
-import { LimiterService } from '../src/types/main.js'
+import { LimiterService } from '../src/types.js'
 
 let limiter: LimiterService
 await app.booted(async () => {
